Reimplemented map-search-term select as a filter
Replaced the tweet recolor in the dots view with a filter/redraw.

This fulfills a requirement of the project spec. It also leaves the class
free to use in linking the dots map to the line chart (TODO- still have to
change class to date).

TODO- need to introduce an "all" button so you can get the full map back.

Also fixed a persistance problem with the search term buttons