Finally got a search term filter for map
Also set up to add a date filter very easily. From the data manipulation
and render standpoin. Have to actually add the element in the page.

The map redraw is a bit slow. That might get worse as we add more dates.
Hopefully not much worse. Dataset is only as large as counties * dates