Reimplemented map search-term filter as re-color
Instead of obliterating and redrawing the whole map, I just loop through
our data, select each county in the map, and change the class of the path
element to a new color.

Oddly, this is if anything slower than deleting the svg, then looping
through the dataset and redrawing the whole thing. I don't understand why
that should be.

However, in this method, you do NOT see the map disappear entirely, which
is nice.  Also, if you are zoomed in and centered on a certain area, you
can recolor without loosing that.

Are these benefits worth an extra slight delay? Is there any way to speed
this up?

There is a one-line change in addMapButtons() that restores the original
svg-redraw method.

Ted, I'd be interested in your thoughts