Added date selection to chloropleth map.
Updated aggregation script to produce a dates json for drawing the
date-selection line.

We want to filter both by date AND by term, dynamically and in concert.
Needed a way to store both values where they'd be globally accessible.
Decided to pack them in a hidden text div.

Finally, updated the event handler function to color the map or redraw the
bar chart, based on a parameter.

Ted, this involved altering your ted2.js  Hope that doesn't mess you up.