Much restyling of page.
Changed span12 container div to be large enough to accomodate map.
However, this was causing it to be moved around when the window is small,
so I absolute positioned it next to the tabs. This caused the tooltips to
show up wrong, so I redid their offsests to fix.

Also changed some HTML such that Bootstrap layouts would move buttons back
to a better location.