Updated AskExplorConc+added opacity to scatterplot
I updated the ask, explore, conclude text for all views. I left the
conclusions blank for the time tab, since this is still under dev. Ted, do
you agree with these conclusions? Do you like the text? Is it too wordy?

RE scatterplot:
So many counties were piling on top of one another in the scatter plot that it was hard to tell if there were cryptic trends being hidden by overlying outliers. To get a sense of the density of counties within the plot, I made individual
circles semi-transparent. I think  this makes the true structure of the
data more clear. What do you think of this Ted?

Finally, I fixed a very small hover-over bug with the 'all_dates' selector
in the map. It wasn't changing the mouse to the hand, and now it does.